MathHelper: Fix rounding to integers...

+++ src/main/java/de/static_interface/reallifeplugin/MathHelper.java
@@ -21,6 +21,6 @@ public class MathHelper
 {
     public static double round(double value)
     {
-        return (int) Math.round(value * 100) / (double) 100;
+        return (double) Math.round(value * 100) / (double) 100;
     }
 }
